predictit/requirements.txt
SQLAlchemy==1.3.6
arch==4.8.1
pandas==0.24.2
pyodbc==4.0.26
Keras==2.2.4
sklearn_extensions==0.0.2
tensorflow==1.13.1
prettytable==0.7.2
matplotlib==3.0.3
plotly==3.10.0
cufflinks==0.16
numpy==1.16.3
scipy==1.2.1
seaborn==0.9.0
statsmodels==0.9.0
lightgbm==2.2.3
pmdarima==1.2.1
scikit_learn==0.21.3
tpot==0.10.2
xgboost==0.90
